LayoutTests/http/tests/workers/service/basic-register-expected.txt

PASS: No service worker is initially registered for this origin
PASS: No service worker is initially registered for this origin in private session
Registered!
PASS: registration object's scope is valid
PASS: registration object's updateViaCache is valid
PASS: A service worker is now registered for this origin
PASS: No service worker is registered for this origin in private session
PASS: registering a script which does not exist rejected the promise
